os/board/rtl8730e: Fix Peripheral related Build Warning and SVACE
- Deleted the unused API and sources file that caused Build Warning and SVACE
Build Warning
- Added return for the function does return
- Removed the unused variable
- Performed the correct casting conversion
- Solved the shadowed declaration without use the same variable name for the global and local
- Added the define/undefine condition checking for the variable
SVACE
- DREF_OF_NULL.ASSIGN : Performed NULL checking
- SIMILAR_BRANCHES : Removed the idetical branches and keep one only
- NO_CAST.INTEGER_DIVISION & INTEGER_OVERFLOW : Performed the correct casting conversion
- Added the comment for unable solve UNREACHEABLE_CODE

+++ os/board/rtl8730e/src/component/mbed/targets/hal/rtl8730e/pwmout_api.c
@@ -132,7 +132,7 @@ void pwmout_init(pwmout_t *obj, PinName pin)
 
 	obj->pwm_idx = pwm_chan;
 	obj->period = 0x10000 * (prescaler + 1) / 40;
-	obj->pulse = 0x1000 * (prescaler + 1) / 40;
+	obj->pulse = 0x1000 * (float)((prescaler + 1)) / 40;
 
 	if (!timer8_start) {
 		pwmout_timer8_init(obj);
